refactor: extract CategoryCardRenderer from MainActivity
